backend/requirements.txt
nltk>=3.7
requests>=2.28.0
python-dotenv>=0.20.0
openai-whisper>=20230918
moviepy>=1.0.3
flask==3.1.0
flask-cors==5.0.1